resolv: Ensure mDNS process can be stopped during probing
The mDNS process might not get stopped in case where it was waiting to start
the mDNS probe. The problem comes from an insecure use of
PROCESS_WAIT_EVENT_UNTIL. Since the mDNS name collision check relies on staring
and stopping the mDNS process this might lead to wrong behaviour of the mDNS
probing.